# openapi/schema.yaml
openapi: 3.0.3
info:
  title: PharmaPocket API
  version: 1.0.0
  description: Contrat de l'API v1 consommée par le frontend PharmaPocket.
paths:
...
  /api/v1/content/landing/:
    get:
      operationId: content_landing
      tags:
      - content
      security:
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LandingPayload'
          description: ''
...
  /api/v1/content/read-state/:
    get:
      operationId: content_read_state_list
      parameters:
      - in: query
        name: slugs
        schema:
          type: string
        description: Slugs séparés par des virgules.
        required: true
      tags:
      - content
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadStateMap'
          description: ''
    post:
      operationId: content_read_state_update
      tags:
      - content
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MicroArticleReadState'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/MicroArticleReadState'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MicroArticleReadState'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadState'
          description: ''
  /api/v1/content/saved/:
    get:
      operationId: content_saved_list
      tags:
      - content
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MicroArticleListItem'
          description: ''
    post:
      operationId: content_saved_create
      tags:
      - content
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavedMicroArticleCreate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SavedMicroArticleCreate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SavedMicroArticleCreate'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedState'
          description: ''
  /api/v1/content/saved/{slug}/:
    get:
      operationId: content_saved_status
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - content
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedState'
          description: ''
    delete:
      operationId: content_saved_delete
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - content
      security:
      - cookieAuth: []
      responses:
        '204':
          description: No response body
...
  /api/v1/tags/:
    get:
      operationId: tag_list
      parameters:
      - in: query
        name: limit
        schema:
          type: integer
      - in: query
        name: q
        schema:
          type: string
      tags:
      - tags
      security:
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagPayload'
          description: ''
...
components:
  schemas:
...
    CardType:
      enum:
      - standard
      - recap
      - detail
      type: string
      description: |-
        * `standard` - standard
        * `recap` - recap
        * `detail` - detail
    CategoryPayload:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
      required:
      - id
      - name
      - slug
...
    ImagePayload:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        url:
          type: string
          nullable: true
        credit_text:
          type: string
          nullable: true
        credit_author:
          type: string
        credit_source:
          type: string
        credit_source_url:
          type: string
        credit_license:
          type: string
        credit_license_url:
          type: string
      required:
      - id
      - title
      - url
    LandingCard:
      type: object
      properties:
        title:
          type: string
        summary:
          type: string
        cta_label:
          type: string
        href:
          type: string
      required:
      - cta_label
      - href
      - summary
      - title
    LandingPayload:
      type: object
      properties:
        title:
          type: string
        hero_title:
          type: string
        hero_subtitle:
          type: string
        hero_bullets:
          type: array
          items:
            type: string
        steps:
          type: array
          items:
            $ref: '#/components/schemas/LandingStep'
        cards:
          type: array
          items:
            $ref: '#/components/schemas/LandingCard'
        primary_cta_label:
          type: string
        primary_cta_target:
          type: string
        secondary_cta_label:
          type: string
        secondary_cta_target:
          type: string
      required:
      - cards
      - hero_bullets
      - hero_subtitle
      - hero_title
      - primary_cta_label
      - primary_cta_target
      - secondary_cta_label
      - secondary_cta_target
      - steps
      - title
...
    LandingStep:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
      required:
      - detail
      - title
...
    MicroArticleListItem:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        answer_express:
          type: string
        takeaway:
          type: string
        key_points:
          type: array
          items:
            type: string
        cover_image_url:
          type: string
          nullable: true
        cover_image_credit:
          type: string
          nullable: true
        cover_image:
          allOf:
          - $ref: '#/components/schemas/ImagePayload'
          nullable: true
        tags:
          type: array
          items:
            type: string
        tags_payload:
          type: array
          items:
            $ref: '#/components/schemas/TagPayload'
        categories_theme_payload:
          type: array
          items:
            $ref: '#/components/schemas/CategoryPayload'
        categories_maladies_payload:
          type: array
          items:
            $ref: '#/components/schemas/CategoryPayload'
        categories_medicament_payload:
          type: array
          items:
            $ref: '#/components/schemas/CategoryPayload'
        categories_pharmacologie_payload:
          type: array
          items:
            $ref: '#/components/schemas/CategoryPayload'
        published_at:
          type: string
          format: date-time
          nullable: true
        decks_count:
          type: integer
          minimum: 0
        card_type:
          $ref: '#/components/schemas/CardType'
        subject:
          allOf:
          - $ref: '#/components/schemas/SubjectSummary'
          nullable: true
      required:
      - answer_express
      - cover_image_url
      - id
      - key_points
      - slug
      - tags
      - takeaway
      - title
    MicroArticleReadState:
      type: object
      description: POST /read-state/ — `validated_data["page"]` est la fiche résolue.
      properties:
        slug:
          type: string
        is_read:
          type: boolean
      required:
      - is_read
      - slug
...
    ReadState:
      type: object
      properties:
        slug:
          type: string
        is_read:
          type: boolean
      required:
      - is_read
      - slug
    ReadStateMap:
      type: object
      properties:
        items:
          type: object
          additionalProperties:
            type: boolean
      required:
      - items
...
    SavedMicroArticleCreate:
      type: object
      description: POST /saved/ — `validated_data["page"]` est la fiche résolue.
      properties:
        slug:
          type: string
      required:
      - slug
    SavedState:
      type: object
      properties:
        saved:
          type: boolean
      required:
      - saved
...
    SubjectSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
      required:
      - description
      - id
      - name
      - slug
    TagPayload:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
      required:
      - id
      - name
      - slug
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid